uint8_t

+++ vblank.h
@@ -5,8 +5,8 @@ class vblank {
 public:
 	vblank(CPU &cpu): _cpu(cpu), _last_int(0), _interrupt(1) {}
 
-	void tick(unsigned long now) {
-		unsigned long dt = now - _last_int;
+	void tick(uint32_t now) {
+		uint32_t dt = now - _last_int;
 		if (dt >= interval) {
 			_cpu.raise(_interrupt);
 			_interrupt = _interrupt == 1? 2: 1;
@@ -17,8 +17,8 @@ public:
 	static const int interval = 1000 / 60 / 2;
 private:
 	CPU &_cpu;
-	unsigned long _last_int;
-	unsigned _interrupt;
+	uint32_t _last_int;
+	uint8_t _interrupt;
 };
 
 #endif
